+++ backend/canisters/group_index/CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+## [[2.0.769](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.769-group_index)] - 2023-07-31
+
 ### Added
 
 - Add API endpoint `lookup_channel_by_group_id` ([#4066](https://github.com/open-chat-labs/open-chat/pull/4066))
+- Handle `set_community_moderation_flags` in `inspect_message` ([#4076](https://github.com/open-chat-labs/open-chat/pull/4076))
 
 ## [[2.0.760](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.760-group_index)] - 2023-07-28